.codex/skills/pdf-to-obsidian-md/pyproject.toml
[project]
name = "pdf-to-obsidian-md"
version = "0.1.0"
description = "PDF to Obsidian Markdown conversion utilities"
requires-python = ">=3.9"
dependencies = [
  "pymupdf>=1.23.0",
  "pillow>=10.0.0",
  "pytesseract>=0.3.10",
]

[project.optional-dependencies]
tables = [
  "pdfplumber>=0.11.0",
  "camelot-py[cv]>=0.11.0",
]
quality = [
  "tqdm>=4.66.0",
  "rich>=13.7.0",
  "rapidfuzz>=3.6.0",
]
test = [
  "pytest>=7.4.0",
  "reportlab>=4.0.0",
]

[tool.pytest.ini_options]
addopts = "-q"
